# crates/ipfrs-interface/openapi.yaml
openapi: 3.0.3
info:
  title: IPFRS HTTP Gateway API
  description: |
    High-performance IPFS-compatible HTTP Gateway and API for IPFRS.

    ## Features
    - Kubo-compatible v0 API endpoints
    - High-speed v1 API with batch operations
    - gRPC interface for advanced use cases
    - WebSocket support for real-time updates
    - Zero-copy tensor API for ML workloads

    ## Authentication
    Most endpoints support optional Bearer token authentication.
    Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-token>
    ```
  version: 0.1.0
  contact:
    name: IPFRS Project
  license:
    name: MIT OR Apache-2.0
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://gateway.ipfrs.io
    description: Production gateway
...
paths:
...
  /v1/tensor/{cid}:
    get:
      summary: Get tensor
      description: Zero-copy tensor retrieval with optional slicing
      tags: [Tensor]
      parameters:
        - name: cid
          in: path
          required: true
          schema:
            type: string
        - name: slice
          in: query
          required: false
          schema:
            type: string
          description: Tensor slice specification (e.g., "0:10,5:15")
      responses:
        '200':
          description: Tensor data
          headers:
            X-Tensor-Shape:
              schema:
                type: string
              example: "[100, 50]"
            X-Tensor-Dtype:
              schema:
                type: string
              example: F32
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/v0/auth/login
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
  - {}